#1, 노드 기본 내장모듈 사용법 및 서버 생성 방법 숙지 & 파일정리 스크립트 개발

+++ Node.js/node-dream/organizeFiles/app.js
@@ -1,7 +1,7 @@
 const fs = require('fs').promises
+const FS = require('fs')
 
 let files = []
-
 
 async function  organizeFiles () {
     // [ 현재 경로에 있는 파일이름, 경로이름 받아오기 ]
@@ -25,25 +25,37 @@ async function  organizeFiles () {
 }
 
 
-function makeForders (forderName) {
-    forderName.forEach(async (name) => {
+async function makeForders (forderName) {
+    for(const name of forderName){
+        if(FS.existsSync(__dirname + `/${name}`)){
+            return
+        }
+
         await fs.mkdir(name)
             .catch(console.error)
-    })
+    }
 }
 
 
-function checkType (files){
-    files.forEach( async (file) => {
+async function checkType (files){
+    for(const file of files){
         const type = file.split(".").pop()
 
         if(type === 'jpg'){
+            if(FS.existsSync(__dirname + `/duplicated/${file}`)){
+                return
+            }
+
             await fs.copyFile(`./${file}`, `./duplicated/${file}`)
                 .catch(console.error)
         }
 
 
-        if(type === 'mp4'){
+        if(type === 'mp4' || type === 'mov'){
+            if(FS.existsSync(__dirname + `/video/${file}`)){
+                return
+            }
+
             await fs.copyFile(`./${file}`, `./video/${file}`)
                 .catch(console.error)
 
@@ -53,19 +65,23 @@ function checkType (files){
 
 
         if(type !== 'jpg' && type !== 'mp4' && file !== 'app.js'){
+            if(FS.existsSync(__dirname + `/captured/${file}`)){
+                return
+            }
+
+
             await fs.copyFile(`./${file}`, `./captured/${file}`)
                 .catch(console.error)
 
             await fs.unlink(`./${file}`)
                 .catch(console.error)
         }
-    })
+    }
 }
 
 
 
 
 organizeFiles()
-
